Fixes sublist exercise by adding missing Relation type.
