+++ exercises/practice/sublist/.meta/config.json
@@ -23,7 +23,8 @@
       ".meta/example.go"
     ],
     "editor": [
-      "cases_test.go"
+      "cases_test.go",
+      "relations.go"
     ]
   }
 }
