+++ exercises/practice/sublist/relations.go
@@ -1,0 +1,12 @@
+package sublist
+
+// Relation is the comparison between lists
+type Relation string
+
+// Possible relations
+const (
+	RelationEqual     Relation = "equal"
+	RelationSublist   Relation = "sublist"
+	RelationSuperlist Relation = "superlist"
+	RelationUnequal   Relation = "unequal"
+)
