+++ exercises/practice/sublist/sublist.go
@@ -1,7 +1,6 @@
 package sublist
 
-// Relation is the comparison between lists
-type Relation string
+// Relation type is defined in relations.go file.
 
 func Sublist(l1, l2 []int) Relation {
 	panic("Please implement the Sublist function")
